[FunctionAttrs] Remove post order "no-recurse" attribute deduction
Summary:
The Attributor can deduce "no-recurse" function attributes since D60074
landed. The support is generally equivalent to the hereby removed
functionality except that we do not erroneously derive "no-recurse" for
potentially redefined functions.

This fixed the bug https://llvm.org/PR41336, where "no-recurse" was
derived for a function with linkonce_odr linkage (which be replaced
later and should therefore not participate in inter-procedural
reasoning).

Test case is available in:
  `llvm/test/Transforms/FunctionAttrs/arg_returned.ll`

Reviewers: homerdin, hfinkel, fedor.sergeev, sanjoy, spatel, nlopes, nicholas, reames

Subscribers: llvm-commits, bollu, hiraditya

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D60075