[Attributor] Iterate until a global fixpoint or a timeout is reached
The Attributor is run multiple times, later with other passes
in-between, until the iteration limit is reached or no progress was
made. This is necessary as (1) not all helper functions use the
optimistic state and we can make progress after a manifest step, and
(2) value created, e.g. promoted, during manifest need to be first
available before we can deduce information for them.

All iteration count changes > 1 should go with changes in the IR, even
if our check lines do not pick them up.