[Attributor][MustExecute] Use optimistic information in MustBeExecutedContextExplorer
Summary:
NOTE: This is a prototype. We need a way to keep
MustBeExecutedContextExplorer updates separate from other updates so
they can continue on, then this should help.

This allows optimistic information to be used to answer the question if
a call returns. This will also record dependences between attributes
using the explorer and AAWillReturn/AANoUnwind that are used to improve
the IR information. If the latter ever get fixed, the updates of the
former are triggered and we have better known information available.

Reviewers: sstefan1, uenoku

Subscribers: hiraditya, bollu, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D68932