[Attr] Introduce the `maxobjsize` attribute
With dereferenceable we track the minimal extend of a pointee but we
lack the ability to track the maximal extend. This is useful as we can
determine alias queries based on the minimal and maximal extend of an
object. Similar to D66157, where we use dereferenceable and exact
allocation information to rule out aliasing (see below), we can use
the maximal extend.

Given two pointers to objects where the objects have respective sizes in
the intervals [a:A] and [b:B], we know that the pointer do not alias,
e.g., the objects are different, if A < b or B < a.

Deduction and use in BasicAA will be provided in follow-up patches.