[Attributor] Add an Attributor CG-SCC pass
Summary:
In addition to the module pass, this patch introduces a CG-SCC pass that
runs the Attributor on a strongly connected component of the call graph
(both old and new PM). The Attributor was always design to be used on a
subset of functions which makes this patch mostly mechanical.

The one change is that we give up `norecurse` deduction in the module
pass in favor of doing it during the CG-SCC pass. This makes an
interfaces simpler but can be revisited.

Reviewers: xbolva00, spatel, chandlerc, uenoku, sstefan1

Subscribers: mehdi_amini, hiraditya, bollu, steven_wu, dexonsmith, dang, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D70767