[OpenMP][WIP] Initial support for `begin/end declare variant`
Summary:
NOTE: This is a WIP patch to foster a discussion. Please do consider
      that when browsing the code. Details will be discussed in
      individual commits once we agreed on the overall model. This is
      also the reason why test coverage, documentation, TODOs, etc. is
      lacking.

This patch provides initial support for `omp begin/end declare variant`,
as defined in OpenMP technical report 8 (TR8).

A major purpose of this patch is to provide proper math.h/cmath support
for OpenMP target offloading. See PR42061, PR42798, PR42799.
The three tests included in this patch show that these bugs (should be)
fixed in this scheme.

In contrast to the declare variant handling we already have, this patch
makes use of the multi-version handling in clang. This is especially
useful as the variants have the same name as the base functions. We
should try to port all OpenMP variant handling to this scheme, see the
TODO in CodeGenModule for a proposed way towards this goal. Other than
that, we tried to reuse the existing multi-version and OpenMP variant
handling as much as possible.

NOTE: There are various TODOs that need fixing and switches that need
      additional cases.

Reviewers: kiranchandramohan, ABataev, RaviNarayanaswamy, gtbercea, grokos, sdmitriev, JonChesterfield, hfinkel, fghanim

Subscribers: rampitec, aheejin, fedor.sergeev, simoncook, bollu, guansong, s.egerton, cfe-commits

Tags: #clang

Differential Revision: https://reviews.llvm.org/D71179