[OpenMP] Use the OpenMPIRBuilder for "omp parallel"
Summary:
NOTE:
This patch needs tests and is missing privatization functionality (see
the TODO). Basic testing was done and any other functionality should be
in place.

This allows to use the OpenMPIRBuilder for parallel regions. Code was
extracted from D61953 and adapted to work with the new version (D70109).

All but one feature should be supported. An update of this patch will
provide test coverage and privatization other than shared.

Reviewers: kiranchandramohan, ABataev, RaviNarayanaswamy, gtbercea, grokos, sdmitriev, JonChesterfield, hfinkel, fghanim

Reviewed By: fghanim

Subscribers: rogfer01, bollu, guansong, cfe-commits

Tags: #clang

Differential Revision: https://reviews.llvm.org/D70290