[OpenMP][WIP] Make the kmp_depend_info type fit in 128 bits.
Summary:
NOTE: The clang codegen change is missing but this simplifies a
      OMPIRBuilder patch and makes the struct more compact. This is not
      to be commited like this!

The old type wasted 64 bits on a length value that is not used. While we
probably need to use the length value eventually, supporting a 64 bit
extend seems excessive. With this patch we have 61 bit extend (=length)
and we use the 3 highest bits for the kind encoding. Alternatively, we
could put some bits in the lower part of the base pointer.

Subscribers: bollu, guansong, openmp-commits

Tags: #openmp

Differential Revision: https://reviews.llvm.org/D71988