[OpenMP][IRBuilder][WIP] Prototype `omp task` support
Summary:
NOTE: The patch is not complete and lacks testing but it should allow
      people to see how the IRBuilder will evolve and give feedback.

This patch adds support for `omp task` to the OpenMPIRBuilder. To
simplify this *and* to make optimizations of tasks possible, we
introduce a new interface: `__kmpc_task`.

Similar to TRegions, the new interface exposes front-end information in
a direct way. More importantly, the task create and task issue step are
conceptually not separated anymore as it is, together with other
non-user related code, hidden inside `__kmpc_task`.

Most of the properties of a task are supported already, incl. dependences.

The outlining logic used for `omp parallel` was (slightly) generalized
and reused but better abstractions are needed.

Reviewers: rogfer01, ABataev, JonChesterfield, kiranchandramohan, fghanim

Subscribers: hiraditya, bollu, guansong, openmp-commits, llvm-commits

Tags: #openmp, #llvm

Differential Revision: https://reviews.llvm.org/D71989