[OpenMP] Avoid crash in preparation for diagnose of unsupported type
Summary:
This was reported as PR45231 but occurs even without `cmath` in C++17
mode if we have an expression involving unsupported types outside of a
function. For now we just avoid to create a diagnose builder in this
case as it might be valid anyway. A proper solution to handle
unsupported types has to be found. Discussion is ongoing in D74387.

Reviewers: hfinkel, ABataev, JonChesterfield

Subscribers: yaxunl, bollu, guansong, ye-luo, openmp-commits, cfe-commits, Fznamznon

Tags: #clang

Differential Revision: https://reviews.llvm.org/D77918