[Attributor][FIX] Handle droppable uses when replacing values
Since we now use the fact that some uses are droppable in the Attributor
we need to handle them explicitly when we replace uses. As an example, a
dead value can as well have live droppable users in which we cannot
replace the value simply by undef. Instead, we either drop the uses (via
`dropDroppableUses`) or keep them as they are. In this patch we do both,
depending on the situation. For values that are dead but not necessarily
removed we keep droppable uses around because they contain information
we might be able to use later. For values that are removed we drop
droppable uses explicitly to avoid replacement with undef.