[Attributor][NFC][WIP] Increase the dependence recompute interval
Summary:
The flag determines how often we throw away dependence information and
start from scratch running all attributes. It does not change the
outcome but only the number of iterations and time it takes to get
there. Initial tests suggest compile time decreases with a higher
number.

TODO: Changes to the tests are missing.

Reviewers: fhahn, lebedev.ri, uenoku, sstefan1

Subscribers: hiraditya, bollu, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D78282