[Attributor][Not-For-Commit] New AAComposeTwoGenericDeduction class
Summary:
The old AAComposeTwoGenericDeduction was removed when the users were
replaced. This new class is superior to the old as it properly handles
generic deductions and their effect on the state. The old usage did not
require this as we only used it with the must-be-executed-context
explorer AA.

This code is not to be committed without use case.

Reviewers: sstefan1, uenoku

Subscribers: hiraditya, bollu, uenoku, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D78718